+++ src/guidelines/patterns/mobile-tab-bar/mobile-tab-bar.specimen.tsx
@@ -1,6 +1,5 @@
 import { useState} from 'react';
-import { MobileTabBar } from '@mattermost/compass-proto';
-import { type MobileTabBarTab } from '@mattermost/compass-proto';
+import { MobileTabBar, type MobileTabBarTab } from '@mattermost/compass-proto';
 import avatarLeonard from '@/assets/avatars/Leonard Riley.png';
 import styles from './mobile-tab-bar.specimen.module.scss';